add basic masking for text format

+++ components/tidb_query_datatype/src/codec/table.rs
@@ -382,7 +382,7 @@ pub fn decode_row_vec(
     data: &mut BytesSlice<'_>,
     ctx: &mut EvalContext,
     cols: &HashMap<i64, ColumnInfo>,
-) -> Result<(Vec<u32>, Vec<Datum>)> {
+) -> Result<(Vec<i64>, Vec<Datum>)> {
     let mut values = datum::decode(data)?;
     if values.get(0).map_or(true, |d| *d == Datum::Null) {
         return Ok((vec![], vec![]));
@@ -403,7 +403,7 @@ pub fn decode_row_vec(
         let v = drain.next().unwrap();
         if let Some(ci) = cols.get(&id) {
             let v = unflatten(ctx, v, ci)?;
-            ids.push(id as u32);
+            ids.push(id);
             datums.push(v);
         }
     }
